Addresses reviewer recommendations
