+++ Indicators/QuantConnect.Indicators.csproj
@@ -183,6 +183,7 @@
     <Compile Include="PythonIndicator.cs" />
     <Compile Include="IntradayVwap.cs" />
     <Compile Include="RelativeVigorIndex.cs" />
+    <Compile Include="RelativeVigorIndexSignal.cs" />
     <Compile Include="SchaffTrendCycle.cs" />
     <Compile Include="WilderMovingAverage.cs" />
     <Compile Include="FractalAdaptiveMovingAverage.cs" />
